+++ src/utils/validationKlasse.ts
@@ -1,0 +1,29 @@
+import { object, string } from 'yup';
+import { toTypedSchema } from '@vee-validate/yup';
+import { DIN_91379A_EXT } from '@/utils/validation';
+import type { TypedSchema } from 'vee-validate';
+
+type ValidationSchema = {
+  selectedSchule: string;
+  selectedKlassenname: string;
+};
+
+export const getValidationSchema = (t: (key: string) => string): TypedSchema<ValidationSchema> => {
+  return toTypedSchema(
+    object({
+      selectedSchule: string().required(t('admin.klasse.rules.schule.required')),
+      selectedKlassenname: string()
+        .matches(DIN_91379A_EXT, t('admin.klasse.rules.klassenname.matches'))
+        .required(t('admin.klasse.rules.klassenname.required')),
+    }),
+  );
+};
+
+export const getVuetifyConfig = (state: {
+  errors: Array<string>;
+}): { props: { error: boolean; 'error-messages': Array<string> } } => ({
+  props: {
+    error: !!state.errors.length,
+    'error-messages': state.errors,
+  },
+});
