+++ src/views/admin/KlassenDetailsView.vue
@@ -15,12 +15,10 @@
   import { type Composer, useI18n } from 'vue-i18n';
   import { useDisplay } from 'vuetify';
   import { type BaseFieldProps, type TypedSchema, useForm } from 'vee-validate';
-  import { toTypedSchema } from '@vee-validate/yup';
-  import { object, string } from 'yup';
-  import { DIN_91379A_EXT } from '@/utils/validation';
   import KlasseForm from '@/components/form/KlasseForm.vue';
   import SuccessTemplate from '@/components/admin/klassen/SuccessTemplate.vue';
-  import KlasseDelete from '@/components/admin/klassen/KlasseDelete.vue'
+  import KlasseDelete from '@/components/admin/klassen/KlasseDelete.vue';
+  import { getValidationSchema, getVuetifyConfig } from '@/utils/validationKlasse';
 
   const route: RouteLocationNormalizedLoaded = useRoute();
   const router: Router = useRouter();
@@ -47,23 +45,11 @@
       : organisationStore.currentOrganisation.name;
   });
 
-  const validationSchema: TypedSchema = toTypedSchema(
-    object({
-      selectedSchule: string().required(t('admin.klasse.rules.schule.required')),
-      selectedKlassenname: string()
-        .matches(DIN_91379A_EXT, t('admin.klasse.rules.klassenname.matches'))
-        .required(t('admin.klasse.rules.klassenname.required')),
-    }),
-  );
+  const validationSchema: TypedSchema = getValidationSchema(t);
 
   const vuetifyConfig = (state: {
     errors: Array<string>;
-  }): { props: { error: boolean; 'error-messages': Array<string> } } => ({
-    props: {
-      error: !!state.errors.length,
-      'error-messages': state.errors,
-    },
-  });
+  }): { props: { error: boolean; 'error-messages': Array<string> } } => getVuetifyConfig(state);
 
   type KlasseEditForm = {
     selectedSchule: string;
@@ -281,7 +267,7 @@
                   md="auto"
                 >
                   <v-btn
-                    class="primary "
+                    class="primary"
                     data-testid="klasse-edit-button"
                     @Click="activateEditing"
                     :block="mdAndDown"
